+++ packages/router/src/index.ts
@@ -14,4 +14,3 @@ export * from './hooks/useLink';
 export * from './hooks/useRoute';
 export * from './hooks/useRouter';
 export * from './navigationFailure';
-export type * from './types/vue-router-alias';